feat: :sparkles: add support for syncthing-macos app install (#167)
As per issue #166 , add in detection of cli on MacOS when using the syncthing-macos package (eg. brew install --cask syncthing ).
This package doesn't add the syncthing cli into the path. Instead it lives hidden under the application package.
When this package is not installed on MacOS, then it falls back to the previous/regular path lookup.


Co-authored-by: jsamuel1 <3156090+jsamuel1@users.noreply.github.com>