feat(syncthing): :sparkles: add ignoring files management
this feature adds a powerful way for the user to modify and manage the ignored files in syncthing
it is managed in the `.stignore` file at the root of the vault.
this will help add ignoring patterns for syncing files
it is currently an early implementation.
still need some testing and refinement on the UI too.